refactor: extract IssueExportRequestContext.swift from IssueExportSupport.swift
Splits the `IssueExportRequestContext` value out. Byte-identical move
of lines 8-13. Both files import Foundation.

IssueExportSupport.swift: 29 → 22 lines. Tests: 667.
