+++ Sources/BugNarrator/Services/IssueExportRequestContext.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct IssueExportRequestContext {
+    let destination: ExportDestination
+    let session: TranscriptSession
+    let selectedIssues: [ExtractedIssue]
+    let apiKey: String
+}
